Added page for the 2024 chooselfie festival
