+++ components/Ranking/Speedline.tsx
@@ -1,0 +1,80 @@
+import { useInfiniteQuery } from "@tanstack/react-query";
+import React from "react";
+
+import { transformSecondsToTimeString } from "@/utils/helperFunctions";
+import useSupabaseBrowser from "@/utils/supabase/client";
+
+import SeeMore from "../SeeMore";
+import { Leaderboard } from "./leaderboard";
+import LoadingSkeleton from "./LoadingSkeleton";
+
+interface Props {
+  highlines_id: string;
+}
+
+const PAGE_SIZE = 5;
+
+function Speedline({ highlines_id }: Props) {
+  const supabase = useSupabaseBrowser();
+
+  async function fetchEntries({ pageParam = 1 }) {
+    const { data, error } = await supabase.rpc("get_crossing_time", {
+      highline_id: highlines_id,
+      page_number: pageParam,
+      page_size: PAGE_SIZE,
+    });
+    return data;
+  }
+
+  const {
+    data: entries,
+    isLoading,
+    isError,
+    fetchNextPage,
+    hasNextPage,
+  } = useInfiniteQuery({
+    queryKey: ["entry", highlines_id, "speedline"],
+    queryFn: ({ pageParam }) => fetchEntries({ pageParam }),
+    initialPageParam: 1,
+    getNextPageParam: (lastPage, pages) => {
+      const nextPage = pages.length + 1;
+      return lastPage?.length === PAGE_SIZE ? nextPage : undefined;
+    },
+    enabled: !!highlines_id.length,
+  });
+
+  if (isLoading) {
+    return <LoadingSkeleton />;
+  }
+
+  if (isError) {
+    return <div>OOPS! Aconteceu algum error. Tente recarregar a página.</div>;
+  }
+
+  return (
+    <>
+      <Leaderboard
+        entries={
+          entries?.pages.flatMap((page, pageIdx) => {
+            return (
+              page?.map((entry, idx) => {
+                if (!entry.crossing_time) return null;
+                return {
+                  name: entry.instagram,
+                  position: pageIdx * PAGE_SIZE + idx + 1,
+                  value: transformSecondsToTimeString(entry.crossing_time),
+                  profilePicture: entry.profile_picture || "",
+                };
+              }) || []
+            );
+          }) || [null, null, null]
+        }
+      />
+      {hasNextPage && (
+        <SeeMore onClick={() => fetchNextPage()} disabled={isLoading} />
+      )}
+    </>
+  );
+}
+
+export default Speedline;
