--- components/tabs/Ranking/LoadingSkeleton.tsx
@@ -1,50 +1,0 @@
-import React from "react";
-
-function LoadingSkeleton() {
-  return (
-    <ul className="mt-2">
-      <li className="mb-2.5 flex flex-row items-center justify-between">
-        <div className="h-4 w-full max-w-[144px] rounded-full bg-gray-200 dark:bg-gray-700" />
-        <div className="h-4 w-full max-w-[112px] rounded-full bg-gray-200 dark:bg-gray-700" />
-      </li>
-      <li className="mb-2.5 flex flex-row items-center justify-between">
-        <div className="h-4 w-full max-w-[144px] rounded-full bg-gray-200 dark:bg-gray-700" />
-        <div className="h-4 w-full max-w-[112px] rounded-full bg-gray-200 dark:bg-gray-700" />
-      </li>
-      <li className="mb-2.5 flex flex-row items-center justify-between">
-        <div className="h-4 w-full max-w-[144px] rounded-full bg-gray-200 dark:bg-gray-700" />
-        <div className="h-4 w-full max-w-[112px] rounded-full bg-gray-200 dark:bg-gray-700" />
-      </li>
-      <li className="mb-2.5 flex flex-row items-center justify-between">
-        <div className="h-4 w-full max-w-[144px] rounded-full bg-gray-200 dark:bg-gray-700" />
-        <div className="h-4 w-full max-w-[112px] rounded-full bg-gray-200 dark:bg-gray-700" />
-      </li>
-      <li className="mb-2.5 flex flex-row items-center justify-between">
-        <div className="h-4 w-full max-w-[144px] rounded-full bg-gray-200 dark:bg-gray-700" />
-        <div className="h-4 w-full max-w-[112px] rounded-full bg-gray-200 dark:bg-gray-700" />
-      </li>
-      <li className="mb-2.5 flex flex-row items-center justify-between">
-        <div className="h-4 w-full max-w-[144px] rounded-full bg-gray-200 dark:bg-gray-700" />
-        <div className="h-4 w-full max-w-[112px] rounded-full bg-gray-200 dark:bg-gray-700" />
-      </li>
-      <li className="mb-2.5 flex flex-row items-center justify-between">
-        <div className="h-4 w-full max-w-[144px] rounded-full bg-gray-200 dark:bg-gray-700" />
-        <div className="h-4 w-full max-w-[112px] rounded-full bg-gray-200 dark:bg-gray-700" />
-      </li>
-      <li className="mb-2.5 flex flex-row items-center justify-between">
-        <div className="h-4 w-full max-w-[144px] rounded-full bg-gray-200 dark:bg-gray-700" />
-        <div className="h-4 w-full max-w-[112px] rounded-full bg-gray-200 dark:bg-gray-700" />
-      </li>
-      <li className="mb-2.5 flex flex-row items-center justify-between">
-        <div className="h-4 w-full max-w-[144px] rounded-full bg-gray-200 dark:bg-gray-700" />
-        <div className="h-4 w-full max-w-[112px] rounded-full bg-gray-200 dark:bg-gray-700" />
-      </li>
-      <li className="mb-2.5 flex flex-row items-center justify-between">
-        <div className="h-4 w-full max-w-[144px] rounded-full bg-gray-200 dark:bg-gray-700" />
-        <div className="h-4 w-full max-w-[112px] rounded-full bg-gray-200 dark:bg-gray-700" />
-      </li>
-    </ul>
-  );
-}
-
-export default LoadingSkeleton;
